[refactor] quiz score를 반영하는 함수를 page로부터 분리
