+++ src/pages/QuizSolvePage/QuizSolvePage.tsx
@@ -3,10 +3,8 @@ import { useEffect } from 'react';
 
 import styled from '@emotion/styled';
 import { Redirect, useHistory } from 'react-router';
-import { v4 } from 'uuid';
 
 import * as QuizServices from '@/api/QuizServices';
-import { updateTotalPoint } from '@/api/UserServices';
 import { POINTS, POST_IDS, USER_ANSWERS } from '@/constants';
 import { useAuthContext } from '@/contexts/AuthContext';
 import { useQuizContext } from '@/contexts/QuizContext';
@@ -14,7 +12,7 @@ import { Layout, QuizContentArea, QuizSubmitArea } from '@components/QuizSolve';
 import useLoading from '@hooks/shared/useLoading';
 import useQuiz from '@hooks/useQuiz';
 
-import type { UserQuizInfo } from '@/interfaces/UserAPI';
+import { updateUserPoint } from './QuizSolvePage.helper';
 
 const QuizSolvePage = () => {
   const history = useHistory();
@@ -31,45 +29,6 @@ const QuizSolvePage = () => {
   } = useQuiz();
 
   const [isLoading, startTransition] = useLoading(true);
-
-  // totalPoint를 받아와 이를 서버에 반영한다.
-  const updateUserPoint = async (totalPoint: number) => {
-    const getNewUserQuizInfo = () => {
-      // userInfo에 들어갈 임시 데이터
-      const newInfo: UserQuizInfo = {
-        _id: v4(),
-        points: totalPoint,
-      };
-
-      if (user.username) {
-        const prevUserInfo = JSON.parse(user.username) as Partial<UserQuizInfo>;
-
-        if (prevUserInfo._id) newInfo._id = prevUserInfo._id;
-
-        if (prevUserInfo.points)
-          newInfo.points = totalPoint + prevUserInfo.points;
-      }
-
-      return newInfo;
-    };
-
-    try {
-      sessionStorage.setItem(POINTS, JSON.stringify(totalPoint));
-
-      // user의 퀴즈 정보가 username에 저장되어 있기 때문에, user.username을 호출한다.
-      // 첫 호출시에는 없을 수 있기 때문에 확인해야 한다.
-
-      // user 정보 업데이트
-      const newUserInfo = await updateTotalPoint({
-        fullName: user.fullName,
-        username: getNewUserQuizInfo(),
-      });
-
-      setUser(newUserInfo);
-    } catch {
-      throw new Error('error occurred at updateUserPoint.');
-    }
-  };
 
   const validate = () => {
     if (quizzes.length !== userAnswers.filter((answer) => answer).length)
@@ -105,7 +64,8 @@ const QuizSolvePage = () => {
     // 로그인했다면, 사용자의 점수를 반영
     if (isAuth) {
       try {
-        await updateUserPoint(totalPoint);
+        const newUserInfo = await updateUserPoint(user, totalPoint);
+        setUser(newUserInfo);
       } catch (error) {
         console.error(error);
       }
